data_collection/requirements.txt
youtube-transcript-api=0.3.1
re
pytube3
